requirements.txt
requests==2.31.0
tenacity==8.2.3
python-dotenv==1.0.0
pytest==7.4.0
psycopg2-binary>=2.9.10dbt-postgres==1.7.0
